stdex::cbitset::operator== made const
This change was made in response to a GCC-14.2.1 warning:
a comparison operator in the form `bool T::operator==(const U&)` (note that the function itself is not const)
expects the left operand to be non-const, but as of C++20 can be "reversed" - so that `T{} == U{}` and `U{} == T{}`
may refer to the same operator overload.
Apparently this may break things.

Regardless of the details regarding this warning, it only makes sense for stdex::cbitset::operator== to allow const
operands on both sides, and it is likely the intended behavior.